fix: support import theme from git repo
